+++ sdk/transit/transit.go
@@ -2,10 +2,7 @@ package transit
 
 import (
 	"fmt"
-	"reflect"
 	"strconv"
-
-	"github.com/gwos/tcg/sdk/logper"
 )
 
 // VersionString defines type of constant
@@ -483,265 +480,6 @@ func (thresholdDescriptor ThresholdDescriptor) String() string {
 	return fmt.Sprintf("[%s, %d]", thresholdDescriptor.Key, thresholdDescriptor.Value)
 }
 
-type BaseInfo struct {
-	// The unique name of the resource
-	Name string `json:"name"`
-	// Type: Required. The resource type of the resource
-	// General Nagios Types are hosts, whereas CloudHub can have richer complexity
-	Type ResourceType `json:"type"`
-	// Owner relationship for associations like hypervisor->virtual machine
-	Owner string `json:"owner,omitempty"`
-	// CloudHub Categorization of resources
-	Category string `json:"category,omitempty"`
-	// Optional description of this resource, such as Nagios notes
-	Description string `json:"description,omitempty"`
-	// Foundation Properties
-	Properties map[string]TypedValue `json:"properties,omitempty"`
-}
-
-func (p *BaseInfo) SetCategory(s string) {
-	p.Category = s
-}
-
-func (p *BaseInfo) SetDescription(s string) {
-	p.Description = s
-}
-
-func (p *BaseInfo) SetName(s string) {
-	p.Name = s
-}
-
-func (p *BaseInfo) SetOwner(s string) {
-	p.Owner = s
-}
-
-func (p *BaseInfo) SetProperty(k string, v interface{}) {
-	t := NewTypedValue(v)
-	if t == nil {
-		logper.Error(nil, "could not set property %s on %s: unsupported value type: %T",
-			k, p.Name, reflect.TypeOf(v))
-		return
-	}
-	if p.Properties == nil {
-		p.Properties = make(map[string]TypedValue)
-	}
-	p.Properties[k] = *t
-}
-
-func (p *BaseInfo) CreateProperties(properties map[string]interface{}) {
-	for k, v := range properties {
-		p.SetProperty(k, v)
-	}
-}
-
-type BaseResource struct {
-	BaseInfo
-	// Device (usually IP address), leave empty if not available, will default to name
-	Device string `json:"device,omitempty"`
-}
-
-func (p *BaseResource) SetDevice(s string) {
-	p.Device = s
-}
-
-// InventoryResource represents a resource that is included in a inventory scan.
-// Examples include:
-//  * nagios host
-//  * virtual machine instance
-//  * RDS database
-//  * storage devices such as disks
-//  * cloud resources such as cloud apps, cloud functions(lambdas)
-//
-// An InventoryResource is the representation of a specific monitored resource during an inventory scan.
-// Each InventoryResource contains list of services (InventoryService) (no metrics are sent).
-type InventoryResource struct {
-	BaseResource
-	// Inventory Service collection
-	Services []InventoryService `json:"services"`
-}
-
-func (p *InventoryResource) AddService(svc InventoryService) {
-	p.Services = append(p.Services, svc)
-}
-
-// String implements Stringer interface
-func (inventoryResource InventoryResource) String() string {
-	return fmt.Sprintf("[%s, %s, %s, %s, %s, %s, %s, %s]",
-		inventoryResource.BaseResource.Name, inventoryResource.BaseResource.Type,
-		inventoryResource.BaseResource.Owner, inventoryResource.BaseResource.Category,
-		inventoryResource.BaseResource.Description, inventoryResource.BaseResource.Device,
-		inventoryResource.BaseResource.Properties, inventoryResource.Services,
-	)
-}
-
-func (p InventoryResource) ToMonitoredResource() MonitoredResource {
-	var services []MonitoredService
-	for _, svc := range p.Services {
-		services = append(services, svc.ToMonitoredService())
-	}
-	return MonitoredResource{
-		BaseResource: p.BaseResource,
-		MonitoredInfo: MonitoredInfo{
-			Status: HostPending,
-		},
-		Services: services,
-	}
-}
-
-// InventoryService represents a Groundwork Service that is included in a inventory scan.
-// In cloud systems, services are usually modeled as a complex metric definition, with each sampled
-// metric variation represented as as single metric time series. During inventory scans, TCG does not gather metric samples.
-//
-// InventoryService collections are attached to an InventoryResource during inventory scans.
-type InventoryService struct {
-	BaseInfo
-}
-
-func (p InventoryService) ToMonitoredService() MonitoredService {
-	return MonitoredService{
-		BaseInfo: p.BaseInfo,
-		MonitoredInfo: MonitoredInfo{
-			Status: ServicePending,
-		},
-		Metrics: []TimeSeries{},
-	}
-}
-
-// String implements Stringer interface
-func (inventoryService InventoryService) String() string {
-	return fmt.Sprintf("[%s, %s, %s, %s, %s, %s]",
-		inventoryService.Name, inventoryService.Type, inventoryService.Owner,
-		inventoryService.Category, inventoryService.Description, inventoryService.Properties,
-	)
-}
-
-type MonitoredInfo struct {
-	// Restrict to a Groundwork Monitor Status
-	Status MonitorStatus `json:"status"`
-	// The last status check time on this resource
-	LastCheckTime *Timestamp `json:"lastCheckTime,omitempty"`
-	// The next status check time on this resource
-	NextCheckTime *Timestamp `json:"nextCheckTime,omitempty"`
-	// Nagios plugin output string
-	LastPluginOutput string `json:"lastPluginOutput,omitempty"`
-}
-
-func (p *MonitoredInfo) SetStatus(s MonitorStatus) {
-	p.Status = s
-}
-
-func (p *MonitoredInfo) SetLastPluginOutput(s string) {
-	p.LastPluginOutput = s
-}
-
-func (p *MonitoredInfo) SetLastCheckTime(t *Timestamp) {
-	p.LastCheckTime = t
-}
-
-func (p *MonitoredInfo) SetNextCheckTime(t *Timestamp) {
-	p.NextCheckTime = t
-}
-
-// A MonitoredResource defines the current status and services of a resource during a metrics scan.
-// Examples include:
-//  * nagios host
-//  * virtual machine instance
-//  * RDS database
-//  * storage devices such as disks
-//  * cloud resources such as cloud apps, cloud functions(lambdas)
-//
-// A MonitoredResource is the representation of a specific monitored resource during a metric scan.
-// Each MonitoredResource contains list of services (MonitoredService). A MonitoredResource does not have metrics,
-// only services.
-type MonitoredResource struct {
-	BaseResource
-	MonitoredInfo
-	// Services state collection
-	Services []MonitoredService `json:"services"`
-}
-
-// String implements Stringer interface
-func (monitoredResource MonitoredResource) String() string {
-	return fmt.Sprintf("[%s, %s, %s, %s, %s, %s, %s, %s, %s]",
-		monitoredResource.BaseResource.Name,
-		monitoredResource.BaseResource.Type,
-		monitoredResource.BaseResource.Owner,
-		monitoredResource.Status,
-		monitoredResource.LastCheckTime,
-		monitoredResource.NextCheckTime,
-		monitoredResource.LastPluginOutput,
-		monitoredResource.BaseResource.Properties,
-		monitoredResource.Services,
-	)
-}
-
-func (p *MonitoredResource) AddService(svc MonitoredService) {
-	p.Services = append(p.Services, svc)
-}
-
-func (monitoredResource *MonitoredResource) ToMonitoredResourceRef() MonitoredResourceRef {
-	return MonitoredResourceRef{Name: monitoredResource.Name, Type: ResourceTypeHost, Owner: monitoredResource.Owner}
-}
-
-func (p MonitoredResource) ToInventoryResource() InventoryResource {
-	var services []InventoryService
-	for _, svc := range p.Services {
-		services = append(services, svc.ToInventoryService())
-	}
-	return InventoryResource{
-		BaseResource: p.BaseResource,
-		Services:     services,
-	}
-}
-
-// A MonitoredService represents a Groundwork Service creating during a metrics scan.
-// In cloud systems, services are usually modeled as a complex metric definition, with each sampled
-// metric variation represented as as single metric time series.
-//
-// A MonitoredService contains a collection of TimeSeries Metrics.
-// MonitoredService collections are attached to a MonitoredResource during a metrics scan.
-type MonitoredService struct {
-	BaseInfo
-	MonitoredInfo
-	// metrics
-	Metrics []TimeSeries `json:"metrics"`
-}
-
-// String implements Stringer interface
-func (monitoredService MonitoredService) String() string {
-	return fmt.Sprintf("[%s, %s, %s, %s, %s, %s, %s, %s, %s]",
-		monitoredService.Name, monitoredService.Type, monitoredService.Owner, monitoredService.Status,
-		monitoredService.LastCheckTime, monitoredService.NextCheckTime,
-		monitoredService.LastPluginOutput, monitoredService.Properties, monitoredService.Metrics,
-	)
-}
-
-func (p MonitoredService) ToInventoryService() InventoryService {
-	return InventoryService{
-		BaseInfo: p.BaseInfo,
-	}
-}
-
-// MonitoredResourceRef references a MonitoredResource in a group collection
-type MonitoredResourceRef struct {
-	// The unique name of the resource
-	Name string `json:"name"`
-	// Type: Optional. The resource type uniquely defining the resource type
-	// General Nagios Types are host and service, whereas CloudHub can have richer complexity
-	Type ResourceType `json:"type,omitempty"`
-	// Owner relationship for associations like host->service
-	Owner string `json:"owner,omitempty"`
-}
-
-// String implements Stringer interface
-func (monitoredResourceRef MonitoredResourceRef) String() string {
-	return fmt.Sprintf("[%s, %s, %s]",
-		monitoredResourceRef.Name,
-		monitoredResourceRef.Type,
-		monitoredResourceRef.Owner,
-	)
-}
-
 // TracerContext describes a Transit call
 type TracerContext struct {
 	AppType    string        `json:"appType"`
@@ -792,59 +530,6 @@ func (operationResults OperationResults) String() string {
 	return fmt.Sprintf("[%d, %d, %s, %s, %d, %d, %s]",
 		operationResults.ResourcesAdded, operationResults.ResourcesDeleted, operationResults.EntityType,
 		operationResults.Operation, operationResults.Warning, operationResults.Count, *operationResults.Results,
-	)
-}
-
-// ResourceGroup defines group entity
-type ResourceGroup struct {
-	GroupName   string                 `json:"groupName"`
-	Type        GroupType              `json:"type"`
-	Description string                 `json:"description,omitempty"`
-	Resources   []MonitoredResourceRef `json:"resources"`
-}
-
-// String implements Stringer interface
-func (resourceGroup ResourceGroup) String() string {
-	return fmt.Sprintf("[%s, %s, %s, %s]",
-		resourceGroup.GroupName, resourceGroup.Type, resourceGroup.Description, resourceGroup.Resources,
-	)
-}
-
-// ResourcesWithServicesRequest defines SendResourcesWithMetrics payload
-type ResourcesWithServicesRequest struct {
-	Context   *TracerContext      `json:"context,omitempty"`
-	Resources []MonitoredResource `json:"resources"`
-	Groups    []ResourceGroup     `json:"groups,omitempty"`
-}
-
-func (p *ResourcesWithServicesRequest) AddResource(res MonitoredResource) {
-	p.Resources = append(p.Resources, res)
-}
-
-// String implements Stringer interface
-func (resourcesWithServicesRequest ResourcesWithServicesRequest) String() string {
-	return fmt.Sprintf("[%s, %s]",
-		resourcesWithServicesRequest.Context.String(),
-		resourcesWithServicesRequest.Resources,
-	)
-}
-
-// InventoryRequest defines SynchronizeInventory payload
-type InventoryRequest struct {
-	Context       *TracerContext      `json:"context,omitempty"`
-	OwnershipType HostOwnershipType   `json:"ownershipType,omitempty"`
-	Resources     []InventoryResource `json:"resources"`
-	Groups        []ResourceGroup     `json:"groups,omitempty"`
-}
-
-func (p *InventoryRequest) AddResource(res InventoryResource) {
-	p.Resources = append(p.Resources, res)
-}
-
-// String implements Stringer interface
-func (inventoryRequest InventoryRequest) String() string {
-	return fmt.Sprintf("[%s, %s, %s]",
-		inventoryRequest.Context.String(), inventoryRequest.Resources, inventoryRequest.Groups,
 	)
 }
